Подключил webpack
