+++ db.json
@@ -57,6 +57,11 @@
       "name": "Илья Панин",
       "phone": "+7 (960) 888-85-25",
       "id": 7
+    },
+    {
+      "name": "Илья Панин",
+      "phone": "+7 (960) 888-85-25",
+      "id": 8
     }
   ]
 }
