+++ index.html
@@ -264,6 +264,6 @@
     </div>
 
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <script src="js/script.js" charset="utf-8"></script>
+    <script src="js/bundle.js" charset="utf-8"></script>
 </body>
 </html>
