+++ package.json
@@ -19,6 +19,8 @@
   "devDependencies": {
     "i": "^0.3.6",
     "json-server": "^0.16.3",
-    "npm": "^7.15.0"
-  }
+    "npm": "^7.15.0",
+    "webpack-cli": "^4.7.2"
+  },
+  "keywords": []
 }
